# fix: Optimizes OPL using string interpolation