--- a/Projects/Server/Items/Item.cs
+++ b/Projects/Server/Items/Item.cs
@@ -1874,35 +1874,35 @@ namespace Server
 
             if (v != 0)
             {
-                list.Add(1060448, v.ToString()); // physical resist ~1_val~%
+                list.Add(1060448, $"{v}"); // physical resist ~1_val~%
             }
 
             v = FireResistance;
 
             if (v != 0)
             {
-                list.Add(1060447, v.ToString()); // fire resist ~1_val~%
+                list.Add(1060447, $"{v}"); // fire resist ~1_val~%
             }
 
             v = ColdResistance;
 
             if (v != 0)
             {
-                list.Add(1060445, v.ToString()); // cold resist ~1_val~%
+                list.Add(1060445, $"{v}"); // cold resist ~1_val~%
             }
 
             v = PoisonResistance;
 
             if (v != 0)
             {
-                list.Add(1060449, v.ToString()); // poison resist ~1_val~%
+                list.Add(1060449, $"{v}"); // poison resist ~1_val~%
             }
 
             v = EnergyResistance;
 
             if (v != 0)
             {
-                list.Add(1060446, v.ToString()); // energy resist ~1_val~%
+                list.Add(1060446, $"{v}"); // energy resist ~1_val~%
             }
         }
 
@@ -1912,15 +1912,7 @@ namespace Server
         public virtual void AddWeightProperty(ObjectPropertyList list)
         {
             var weight = PileWeight + TotalWeight;
-
-            if (weight == 1)
-            {
-                list.Add(1072788, weight.ToString()); // Weight: ~1_WEIGHT~ stone
-            }
-            else
-            {
-                list.Add(1072789, weight.ToString()); // Weight: ~1_WEIGHT~ stones
-            }
+            list.Add(weight == 1 ? 1072788 : 1072789, $"{weight}");
         }
 
         /// <summary>