--- a/Projects/Server/PropertyList/IPropertyList.cs
+++ b/Projects/Server/PropertyList/IPropertyList.cs
@@ -1,0 +1,9 @@
+namespace Server;
+
+public interface IPropertyList
+{
+    public void Terminate();
+    // TODO: Use string interpolator
+    public void Add(int number, string arguments = null);
+    public void Add(string text);
+}