--- a/Projects/Server/PropertyList/IPropertyList.cs
+++ b/Projects/Server/PropertyList/IPropertyList.cs
@@ -17,7 +17,7 @@ using Server.Text;
 
 namespace Server;
 
-public interface IPropertyList : ISelfInterpolatedStringHandler<IPropertyList>
+public interface IPropertyList : ISelfInterpolatedStringHandler
 {
     public void Reset();
     public void Terminate();