--- a/Projects/Server/PropertyList/PropertyListStringInterpolator.cs
+++ b/Projects/Server/PropertyList/PropertyListStringInterpolator.cs
@@ -1,0 +1,3 @@
+namespace Server;
+
+