--- a/Projects/Server/PropertyList/PropertyListStringInterpolator.cs
+++ b/Projects/Server/PropertyList/PropertyListStringInterpolator.cs
@@ -1,3 +1,0 @@
-namespace Server;
-
-