… Projects/Server/ObjectPropertyList.cs
… Projects/Server/PropertyList/ObjectPropertyList.cs
… Projects/Server/Text/ISelfInterpolatedStringHandler.cs
… Projects/Server/Text/StringHelpers.cs
… Projects/Server/Utilities/Utility.cs
… Projects/UOContent/Engines/Bulk Orders/LargeBOD.cs
… Projects/UOContent/Engines/Bulk Orders/SmallBOD.cs
… Projects/UOContent/Engines/CannedEvil/ChampionSpawn.cs
… Projects/UOContent/Engines/ConPVP/Gumps/AcceptTeamGump.cs
… Projects/UOContent/Engines/ConPVP/Gumps/ConfirmSignupGump.cs
… Projects/UOContent/Engines/ConPVP/Gumps/TournamentBracketGump.cs
… Projects/UOContent/Engines/Factions/Items/Sigil.cs
… Projects/UOContent/Engines/Factions/Items/Traps/FactionTrapRemovalKit.cs
… Projects/UOContent/Engines/Factions/Mobiles/Guards/BaseFactionGuard.cs
… Projects/UOContent/Engines/ML Quests/Items/APersonalLetterAddressedToAhie.cs
… Projects/UOContent/Engines/ML Quests/Items/AlchemistsBandage.cs
… Projects/UOContent/Engines/ML Quests/Items/BasinOfCrystalClearWater.cs
… Projects/UOContent/Engines/ML Quests/Items/BridesLetter.cs
… Projects/UOContent/Engines/ML Quests/Items/CompletedTuitionReimbursementForm.cs
… Projects/UOContent/Engines/ML Quests/Items/CrateForSledge.cs
… Projects/UOContent/Engines/ML Quests/Items/DreadSpiderSilk.cs
… Projects/UOContent/Engines/ML Quests/Items/FragmentOfAMapDelivery.cs
… Projects/UOContent/Engines/ML Quests/Items/FriendsOfTheLibraryApplication.cs
… Projects/UOContent/Engines/ML Quests/Items/GiftForArielle.cs
… Projects/UOContent/Engines/ML Quests/Items/NotarizedApplication.cs
… Projects/UOContent/Engines/ML Quests/Items/OfficialSealingWax.cs
… Projects/UOContent/Engines/ML Quests/Items/PortraitOfTheBride.cs
… Projects/UOContent/Engines/ML Quests/Items/PrismaticAmber.cs
… Projects/UOContent/Engines/ML Quests/Items/QuestGiverItem.cs
… Projects/UOContent/Engines/ML Quests/Items/ReginasLetter.cs